feat(core): enable middleware always, add password protection and conditional redirects
- Rename middleware__DISABLED.ts to middleware.ts (middleware always active)
- Integrate AuthenticationService for password protection flow
- Run redirects only when ENABLE_REDIRECTS_MIDDLEWARE is set at runtime
- Remove middleware renaming logic from CLI generate.ts

Made-with: Cursor
